Moving files into component folder
