+++ src/reduxExample/component/Posts.js
